% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{array_sum}
\alias{array_sum}
\title{array_sum}
\usage{
array_sum(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the sum of the vector
}
\description{
Given an vector of integers, find the sum of its elements without using sum() function
}
\examples{
x = [1, 2, 3], so return 6
}
