% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{init_user}
\alias{init_user}
\title{My hello world function}
\usage{
init_user(x)
}
\arguments{
\item{x}{The name of the person}
}
\value{
The output from \code{\link{print}}
}
\description{
My hello world function
}
\examples{
init_user("Sam")


}
