% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{min_max}
\alias{min_max}
\title{Write a function that finds the minimum and maximum value in a vector without using min() or max()}
\usage{
min_max(x)
}
\arguments{
\item{x}{-a vector}
}
\value{
the minimum and maximum value of a vector
}
\description{
Write a function that finds the minimum and maximum value in a vector without using min() or max()
}
