% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{plus_minus}
\alias{plus_minus}
\title{plus_minus}
\usage{
plus_minus(x)
}
\arguments{
\item{x}{a vector}
}
\value{
decimal value of each fraction on a new line
}
\description{
Given an array of integers, calculate the fraction of its elements that are positive, negative and zero
}
