% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{test}
\alias{test}
\title{My hello world function}
\usage{
test(x)
}
\arguments{
\item{x}{The name of the person}
}
\value{
The output from \code{\link{print}}
}
\description{
My hello world function
}
\examples{
test("Sam")

}
